The Digital Yearbook
Remember how we used to get our classmates to sign the pages of our physical yearbooks for us?

People would write nice things about each other, make predications about each other's futures, and mention shared experiences to ensure they were recoverable if forgotten. The reason that schools do this is to reinforce the sense of community between people in a way that lasts. Alumni stay in touch better, donate more, and generally think more highly of their schools. Many cherish their old yearbooks and pull them out to look at them from time to time. It was a good practice to reinforce positive culture in a community. Digital transformation has changed this practice, and it never really applied to organizations additional to schools to begin with (can you think of a company that had a yearbook?). I think there's an opportunity to bring it back digitally in a meaningful way within communities generally. 

This app allows anyone with a list of email addresses and names to begin a series of chain mails soliciting positive comments from a community. For each participant, the app chooses a community member at random and asks the participant to write something nice about that community member. Whenever anyone elects to write something nice about someone else, the note is shared with the community member (and saved in a master spreadsheet). This community member is then shown the message and is asked to write about someone else at random. Community members can elect to write about someone else whenever they want. The more messages they write, the more likely they are to be selected as the recipient of a comment from someone else. Community members have the option to opt out of the digital yearbook. 

This app is built in Google Script (JavaScript) and relies on Google Sheets and Google Forms to function. It also relies on the members of the community to be in a Google Group together so that the form can automatically collect email addresses and prohibit non-members from joining. It is not hard to change the script to enable any type of email address to work (disabling the auto-email collection and security feature), but it will be more difficult to remove it from the Google Sheets and Google Forms that it relies on currently.